Add message/event streaming code
This change adds functionality that layers decoding of JSON based
messages on top of a WebSocket stream. Conceptually it is very similar
to Tokio's concept of a Framed stream, but this code is also aware of
WebSocket specific control messages such as pings (to which it
transparently responds) and closes (which gracefully stop the stream).