Use tracing crate for providing tracing functionality
Many of our crates that use async & await syntax have migrated to using
the tracing crate instead of regular logging using log.
With this change we do the same for this crate. By going down this route
we are able to attach additional context to events emitted. In addition
to switching over we also add a few more trace points which are of
potential interest.