Send repeated ping messages and expect pongs
During an attempt to stream data the Internet connection dropped.
Unfortunately, such a loss of connectivity did not result in a detected
broken connection as we envisioned, but just stopped active streaming of
messages.
In order to detect such problems we need some active entity that will
detect issues of this nature, instead of relying on the tungstenite
crate (which seemingly is just a passive decoder). To that end, this
change introduces a facility issuing pings to the server we are
connected to in regular intervals. If the server repeatedly fails to
respond to pings, we will error out and close the connection on our end.