Simplify streaming functionality
The stream function and all the functionality it relies upon has never
been a shiny example of well crafted code. However, with the
introduction of the ping functionality complexity arguably got out of
hand.
This change marks a first step in the process of untangling the code a
bit, reducing the complexity in the process. It essentially inlines the
previously existent decode_msg function directly into handle_msg. It
also acknowledges the fact that what are not really interested in a
"pong" message from the server per se, as long as there is any message
coming from it. Given those two changes, we can remove three of the five
variants of our Operation enum.