Remove Operation enum
Now that we have reduced the number of variants that the Operation enum
needs to contain we can see much better that it really is just a custom
version of an Option.
Thus, with this patch we remove the enum altogether, replacing it with
an option. We also "transpose" part of the return type of the handle_msg
function, which makes it easier for downstream code to work with the
result.