Do not send pings if we recently received a message anyway
With the initial version of the ping functionality we just
unconditionally sent a ping message to the server whenever an interval
elapsed. However, in many cases we may have just received some kind of
message that already indicates to us that the connection is still alive.

This patch improves the ping logic in such a way that we no longer send
unnecessary pings to the server if we have received any kind of message
recently.