Publicly export tungstenite crate
Both the tokio-tungstenite and async-tungstenite crates have public
exports of the underlying tungstenite crate. Doing so potentially makes
it easier for client code to access underlying functionality. With this
change we do the same. The main benefit is that now client code no
longer requires an explicit dependency on tungstenite and will
automatically get migrated over to using the version that this crate
uses.