Use our WebSocketStream in more locations
For testing purposes we maintain our own WebSocketStream typedef.
The serve_and_connect test function returns some unwieldy monstrosity of
a type, but really should use this typedef.
With this change we adjust the typedef and our mock_server function so
in such a way that all types to match up and we can use WebSocketStream
everywhere.