Introduce Wrapper type
This change introduces the Wrapper type. As the name tries to convey,
this type represents a wrapper around a "raw" websocket stream as
established via tungstenite. The wrapper fulfills the same purpose as
our existing stream() function: it transparently takes care of handling
control messages such as pongs and initiates the sending of pings at
certain user-defined intervals. As such, it constitutes a complete
reimplementation of the existing functionality.
Instead of yielding vectors of bytes, the new implementation works with
Message objects, which distinguish between text and binary data -- just
like tungstenite and the websocket protocol itself does.
Furthermore, this type is built as a large state machine and not using
the various future based combinators. This should allow it to be more
efficient in execution speed but also have a positive effect on crate
compilation times.
Lastly, because we no longer build the logic based on the unfold
function along with various combinators fudged on top, but rather offer
a wrapper struct as the interface of choice, we can also implement not
only the Stream trait but also Sink, which makes the crate more
versatile. Not only that, we now provide a nameable type to users, which
is in contrast to the return type of the stream() function, which was
opaque.