Make wrap module public
In the future we intend to add more websocket related functionality to
the crate. As such it makes sense to introduce further structure to the
public API surface.
With this change we make the wrap module that we introduced earlier
public and remove the root-level re-exports that we used from it earlier
from the crate's exports.