Allow for no pings to be sent
By default we configure our Wrapper in such a way that it sends ping
messages to the server every 30s. With this change we allow users to
completely disable the sending of pings. While at it, this change adds a
test to verify that pings can be disabled and also another one checking
that pings are actually sent when the feature is enabled.