Send no pongs by default
As it turns out the default tungstenite websocket client already
responds to ping messages automatically with pongs. As such, the custom
logic that we have for sending pongs just causes duplicate messages to
be sent.
Hence, with this change we disable the sending of pong messages in
response to pings by default. We still allow the user to enable the
feature on demand, via the newly introduced Builder::set_send_pongs
method.