Enhance tracing for binary messages
The trace points we recently added log, among other things, websocket
messages sent and received. Those messages can reasonably contain binary
data, but sometimes they may be "binary" messages that actually contain
clear text content such as JSON. Currently such messages would be
printed as a dump of decimal bytes.
This representation makes it cumbersome to understand what is going on
and so with this change we enhance the logic in such a way that we try
to convert binary messages to a valid string and log that. Only if that
fails do we fall back to logging a byte dump.