Run stress tests in CI
With the rewrite of the streaming functionality we have decreased
reported code coverage significantly. One of the reasons, as evidenced
by the coverage report, is that we added a stress test that is not
actually run in CI currently.
There is no really good reason for not running it there and so with this
change we execute it both as part of the test run itself and for
coverage collection.