Update minimum supported Rust version to 1.53.0
With upcoming changes we would like to make use of features only
available in newer versions of Rust. To make this possible this change
updates the minimum supported version of Rust to 1.53.0.
Note that because this change is only required for tests at this point,
we do not treat it as compatibility breaking and aim to potentially
publish another minor patch bump.