Introduce subscription functionality
This change introduces a new module to the crate, subscribe. The module
provides fairly generic way to create some kind of subscription object
in client code: basically an entity associated with a stream of sorts
that can control certain behavior pertaining said stream at runtime and
without acquiring a mutable reference to it. That is how, conceptually,
the historical aggregate streaming in many Brokerage APIs works, where
one can adjust subscription at runtime, by sending some kind of control
message over the stream and then waiting for a response in-line.

At the implementation level we introduce two main types: MessageStream
represents the stream of message from a websocket server and a
Subscription object is associated with such a stream and allows for
changing its state. A MessageStream then basically acts as a router
where all messages classified as "control message" are forwarded to the
associated Subscription and everything else is emitted through the
stream directly. Such classification is user controlled: we introduce
the Message trait which is used to classify an incoming message as
either a control message or "user message".
One implication of this design is that, in order for a Subscription to
ever receive a control message response, the respective MessageStream
needs to be polled -- otherwise control messages would not be received
and, hence, never be forwarded to a Subscription.
In such a scheme awareness for errors is unavoidable. A websocket
stream, for example, can always fail. Similarly, higher level messages
sent over it often have to be decoded form a byte representation first,
which almost necessarily is a fallible operation.
By virtue of how MessageStream and Subscription objects interact with
each other, we need to take certain errors into account. Specifically,
when a Subscription sends a control message over the associated channel,
it waits for a control response from the upstream entity (e.g., a
websocket server). If we fail to decode said message, we should unblock
the operation and report an error, as it is unlikely that more control
messages arrive. We piggy-back on the aforementioned classification step
via the Message trait to infer whether a message constitutes an error or
not, because in general that is knowledge that is specific to a concrete
client. Any message classified as error this way then causes an error to
be sent to the Subscription object, which in turn will make sure to emit
this error for a control operation in progress.