Introduce drive helper function for resolving subscription futures
In order to control a websocket channel through a Subscription object
the associated stream needs to be polled. That is necessary because the
stream effectively acts as a splitter: it extracts control messages and
forwards them to the corresponding subscription, while merely passing
through non-control messages.
Typically this is not a problem, except if a user has not yet started
processing the stream. This change introduces a helper method that can
be used to "drive" a Subscription related future to conclusion, making
sure that the corresponding stream is polled.