Add more tests pertaining error awareness
Awareness of errors is a basic requirement for our MessageStream and
Subscription types. Specifically, the Subscription should stop control
operations of the underlying stream reported an error of sorts.
To make sure that we are handling such cases properly, this change adds
more test cases that exercise the error paths.