Introduce Subscription::read method
Sometimes a protocol requires that a message be read without requiring a
send happening first. E.g., an API may send a "connected" message after
a connection has been established and this message would need to be
consumed before reading additional messages.
To cater to these use cases, this change introduces the
Subscription::read method. This method is similar to the existing send,
in that it installs the one-shot channel etc., but it does not require
sending of a message first.