Use unpin helpers
Emulating our own "pin projections", we create a bunch of new Pin
objects in our code. As it turns out that is largely unnecessary given
that we assume that our internally used stream/sink are Unpin and that
there are a bunch of helpers that provide the needed functionality if
that is the case.
To that end, this change simplifies the code a bit by using these unpin
helper methods.