Run clippy using the stable toolchain
So far we were running clippy on our minimum supported Rust version in
CI. Doing so is fine conceptually, but it means one more version to
update. What's more, we also don't get the latest and greatest
suggestions that clippy provides.
With this change we switch to running clippy on the most recent stable
toolchain, which fixes these problems.