Stop emitting debug information in CI
There is not really a good reason to have our CI debug builds emit debug
information in the created binaries. All that really does is bloat said
binaries (which comes along increases in compile times).
With this change we disable creation of debug information. The only
impact this would have on debugability in CI is that file names and line
numbers may be missing in backtraces. We currently do not print
backtraces anyway, so this is not a huge concern.