Overhaul Test CI workflow
This change overhauls the Test CI workflow in a few ways. First and
foremost, we move testing and coverage collection out of the matrix. The
matrix is now restricted to only building. Second, said matrix now
builds with both the dev and release profile. Lastly, we add an explicit
Rust doc run, as that is not currently done by cargo-llvm-cov.