Use io::ErrorKind::TimedOut for ping timeouts
When we miss receipt of a pong from the server in response to a ping
that we sent, we report an error over the stream. Currently this error
use the std::io::ErrorKind::Other variant.
Admittedly, the std::io::ErrorKind::TimedOut variant is more fitting and
to the point and so with this change we switch over to using it instead.