Add test for missed interval behavior
It is conceivable that our Wrapper type skips sending a ping if the
corresponding object is not polled. In the past (i.e., before commit
754af0396d77 ("Fix ping sending after change in tokio Interval API"),
that would have led to a burst of wake ups, which would result in missed
timeouts. While this issue has been fixed, this change adds a test for
this scenario.