Add test making sure we sent pings after timeout
If our Wrapper type encounters a timeout for a ping sent, the behavior
we want to provide is that we report an error but then also continue
sending pings. That is so that clients can opt to ignore timeouts
entirely or for a certain number of attempts.
To make sure that this is the behavior we ultimately provide, this
change adjusts the existing test checking the timeout behavior
accordingly.