Properly set waker in Pinger logic
As it turns out our polling logic inside the Pinger type is not quite
idiomatic: when the next_ping.poll_tick() call returns Poll::Ready, no
further wake ups are scheduled automatically. We use the reset() method
to re-arm the interval, but that is arguably not intended use of the
API. On top of that, there is an actual correctness issue: if the waker
associated with the provided context changed, we will keep waking up the
original waker instead of the current one.
This change fixes this problem by invoking next_ping.poll_tick() until
it returns Poll::Pending. Once that happens the current context's waker
is on the hook for being woken up. To preserve the intended "missed
ticks" behavior, we explicitly set that during construction.