Lower tokio dependency to 1.8
We recently bumped the minimum version of tokio to 1.27 owing to a
change in behavior of their Interval API. However, as it turns out we
were actually using the API in an arguably wrong way. That has been
corrected by now and with that change the code works fine on any version
of tokio that includes MissedTickBehavior support. To that end, lower
the required version of the dependency to 1.8.