Preserve send message state when flush is pending
When the SendMessageState::Flush variant is active, we unconditionally
advance the state to Unused, even if the poll_flush_unpin() call returns
Pending. While unlikely to cause issues (because the flush will still go
through eventually), it seems more apt and in line with general future
workings to not advance the state and keep polling the flush until it
truly concluded.