Advance message sending by more than one step at a time
It does not seem proper to advance the sending of messages as
represented by the SendMessageState type with one state transition at a
time. Rather, we should advance it as far as possible, until we either
cannot continue because an IO is pending, an error occurred, or we are
done sending altogether.