Update tokio-tungstenite dependency to 0.26
This change updates the tokio-tungstenite dependency we rely on to
version 0.26. Upstream switched to a more efficient representation of
message contents. However, in the theme of abstracting over their
implementation details, we keep our Message enum the as it has been,
which results in some arguably avoidable allocations. We can revisit
this decision in the future, should it become an issue.